chore: single file & 3.6 import
