+++ setup.cfg
@@ -87,9 +87,9 @@ validate_pyproject.tool_schema =
     setuptools = validate_pyproject.api:load_builtin_plugin
     distutils = validate_pyproject.api:load_builtin_plugin
 repo_review.checks =
-    validate_pyproject = validate_pyproject.repo_review.checks:repo_review_checks
+    validate_pyproject = validate_pyproject.repo_review:repo_review_checks
 repo_review.families =
-    validate_pyproject = validate_pyproject.repo_review.families:repo_review_families
+    validate_pyproject = validate_pyproject.repo_review:repo_review_families
 
 [tool:pytest]
 # Specify command line options as you would do when invoking pytest directly.
